requirements.txt
flask==2.2.2
gunicorn
pytorch==1.13
openai-whisper
pysrt==1.1.2
googletrans==4.0.0-rc1
pytube==15.0.0
backoff
